lerp some playerstate values on clientside, make progress u16 instead of u8 for more precision

+++ source/gameshared/gs_synctypes.h
@@ -515,7 +515,7 @@ struct SyncPlayerState {
 	Team real_team;
 
 	BombProgress progress_type;
-	u8 progress;
+	u16 progress;
 };
 
 enum UserCommandButton : u8 {
